group axios requests togrther
