+++ src/containers/CarsContainer.js
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 import _ from 'lodash';
-import { fetchFavs } from '../redux/favs/favActions';
-import { fetchCars } from '../redux/cars/carActions';
+import { fetchFavs, fetchCars } from '../redux/axiosRequests';
+
 import Toolbar from '../components/Toolbar/Toolbar';
 
 function CarsContainer() {
