+++ src/redux/Store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 
 import thunk from 'redux-thunk';
-import rootReducer from './rootReducer';
+import rootReducer from './reducers/rootReducer';
 
 const Store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
