+++ src/redux/constants/favTypes.js
