+++ src/redux/reducers/rootReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import carReducer from './cars/carReducer';
-import favReducer from './favs/favReducer';
-import userReducer from './user/userReducer';
+import carReducer from './carReducer';
+import favReducer from './favReducer';
+import userReducer from './userReducer';
 
 const rootReducer = combineReducers({
   carList: carReducer,
