+++ src/redux/reducers/userReducer.js
@@ -2,7 +2,7 @@ import {
   FETCH_USER_REQUEST,
   FETCH_USER_SUCCESS,
   FETCH_USER_FAILURE,
-} from './userTypes';
+} from '../constants/userTypes';
 
 const initialState = {
   loading: false,
